Move Row class to separate file
