+++ scraper.rb
@@ -5,48 +5,9 @@ require 'scraperwiki'
 require 'nokogiri'
 require 'open-uri/cached'
 require_relative 'lib/table'
+require_relative 'lib/row'
 
 require 'pry'
-
-class Row
-  def initialize(tds)
-    @tds = tds
-  end
-
-  def to_h
-    {
-      name: name,
-      name__mn: name_mn,
-      party: party,
-      term: term,
-      wikiname: wikiname,
-    }
-  end
-
-  private
-
-  attr_reader :tds
-
-  def name
-    tds[1].xpath('.//a').text.strip
-  end
-
-  def name_mn
-    tds[2].text.strip
-  end
-
-  def party
-    tds[4].text.strip
-  end
-
-  def term
-    '2016'
-  end
-
-  def wikiname
-    tds[1].xpath('.//a[not(@class="new")]/@title').text.strip
-  end
-end
 
 class Page
   def initialize(url)
